+++ charts/warden/values.yaml
@@ -12,10 +12,10 @@ fullnameOverride: ""
 #Service configuration
 global:
   operator:
-    image: europe-docker.pkg.dev/kyma-project/dev/warden/operator:PR-41
+    image: europe-docker.pkg.dev/kyma-project/dev/warden/operator:PR-43
 
   admission:
-    image: europe-docker.pkg.dev/kyma-project/dev/warden/admission:PR-41
+    image: europe-docker.pkg.dev/kyma-project/dev/warden/admission:PR-43
 
   config:
     dir: /workspace
@@ -31,7 +31,7 @@ global:
         systemNamespace: "{{ .Release.Namespace }}"
         serviceName: "{{ .Chart.Name }}-admission"
         secretName: "{{ .Chart.Name }}-admission-cert"
-        timeout: 20s
+        timeout: 10s
         port: 8443
       operator:
         metricsBindAddress: "127.0.0.1:8080"
